requirements.txt
streamlit
myproject-core @ git+https://github.com/YOUR_USERNAME/myproject-library.git